introduce test_slots.py, and move the relevant tests there. The idea is to have at least an individual test for each slot
